refactor: Port to Qt 6

+++ src/blackcore/threadedreader.h
@@ -144,7 +144,7 @@ namespace BlackCore
         int m_initialTime = -1; //!< Initial start delay
         int m_periodicTime = -1; //!< Periodic time after which the task is repeated
         QDateTime m_updateTimestamp; //!< when file/resource was read
-        uint m_contentHash = 0; //!< has of the content given
+        size_t m_contentHash = 0; //!< has of the content given
         std::atomic_bool m_markedAsFailed { false }; //!< marker if reading failed
         bool m_unitTest { false }; //!< mark as unit test
         BlackMisc::Network::CUrlLogList m_urlReadLog; //!< URL based reading can be logged
