+++ src/blackmisc/containerbase.h
@@ -105,7 +105,7 @@ namespace BlackMisc
         }
 
         //! Simplifies composition, returns 0 for performance
-        friend uint qHash(const Derived &) { return 0; } // clazy:exclude=qhash-namespace
+        friend size_t qHash(const Derived &) { return 0; } // clazy:exclude=qhash-namespace
 
         //! \copydoc BlackMisc::Mixin::JsonByMetaClass::toJson
         QJsonObject toJson() const
